Moved water supply test fixture
